+++ packages/connect-evm/src/types.ts
@@ -82,6 +82,11 @@ type AddEthereumChainRequest = {
   params: [AddEthereumChainParameter];
 };
 
+type AccountsRequest = {
+  method: 'eth_accounts' | 'eth_coinbase';
+  params: [];
+};
+
 type GenericProviderRequest = {
   method: Exclude<
     string,
@@ -102,6 +107,7 @@ export type ProviderRequest =
   | RevokePermissionsRequest
   | SwitchEthereumChainRequest
   | AddEthereumChainRequest
+  | AccountsRequest
   | GenericProviderRequest;
 
 export type ProviderRequestInterceptor = (