+++ packages/connect-evm/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added analytics tracking of request handling ([#46](https://github.com/MetaMask/connect-monorepo/pull/46))
 - Initial release ([#21](https://github.com/MetaMask/connect-monorepo/pull/21))
 
 ### Fixed